Some hacking on the debian packager script

+++ package/make-debian-package.pl
@@ -2,23 +2,104 @@
 
 use strict;
 
+my $prefix = '/usr/local/antiweb';
+
+sub usage {
+  print STDERR <<END;
+Antiweb .deb packager script (C) Doug Hoyte
+usage: $0 <lisp> <architecture> <operating system>
+
+Valid lisp parameters:
+  cmucl
+  ccl
+
+Valid architecture parameters:
+  amd64
+  i386
+
+Valid operating system parameters:
+  linux
+  freebsd
+END
+  exit 1;
+}
+
+## SCRIPT PARAMETERS
+
+my $lisp = shift || usage();
+my $arch = shift || usage();
+my $os = shift || usage();
+
+
+## VERIFY SCRIPT PARAMETERS
+
 my $bits = length pack("l!")==8 ? 64 : 32;
 
-my $tag = `git describe --tags --match antiweb-\*`;
+if ($lisp eq 'cmucl') {
+  die "packaging cmucl for 64 bit platforms is not supported by this script (but is possible)" unless $arch eq 'i386';
+} elsif ($lisp eq 'ccl') {
+  die "clozurecl is only stable on 64 bit platforms" unless $bits == 64;
+} else {
+  print STDERR "Invalid lisp parameter: $lisp\n\n";
+  usage();
+}
 
-print "QQQ: $tag\n";
+if ($arch eq 'amd64') {
+  die "this machine isn't an amd64" unless $bits == 64;
+} elsif ($arch eq 'i386') {
+  die "this machine isn't an i386" unless $bits == 32;
+} else {
+  print STDERR "Invalid architecture parameter: $arch\n\n";
+  usage();
+}
 
+if ($os eq 'linux') {
+  die "this machine isn't running linux" unless `uname -a` =~ /linux/i;
+} elsif ($os eq 'freebsd') {
+  die "this machine isn't running freebsd" unless `uname -a` =~ /freebsd/i;
+} else {
+  print STDERR "Invalid operating system parameter: $os\n\n";
+  usage();
+}
+
+
+
+## FIND ANTIWEB VERSION
+
+my $aw_version = `git describe --tags --match antiweb-\*`;
+print "Antiweb version: $aw_version\n";
+
+
+## SOME BASIC SANITY CHECKS
+
+die "This antiweb repo has a ../local.lisp. Please back it up and remove it." if -e '../local.lisp';
+
+
+=pod
+if ($lisp eq 'cmucl') {
+  die "Expected to find a cmucl installation at $prefix/"
+    unless (-d $prefix && -d "$prefix/
+} elsif ($lisp eq 'ccl') {
+}
+=cut
+
+
+
+## CONSTRUCT PACKAGE
 
 die "Unable to find DEBIAN/ directory" unless -d 'DEBIAN';
 
 sys("rm -rf build");
-sys("cp -R DEBIAN build");
+sys("mkdir build");
+
+sys("cp -r DEBIAN/ build");
 
 sys("mkdir -p build/usr/local/antiweb");
 
-sys("mkdir -p build/usr/local/antiweb");
 
 
+
+## UTILITIES
 
 sub sys {
   my $cmd = shift;
